+++ src/token.rs
@@ -25,5 +25,5 @@ pub struct Token {
     pub token: TokenType,
     pub lexeme: String,
     pub literal: Option<AplType>,
-    pub line: i32,
+    pub line: usize,
 }
